fix(cake_diffusion): support CUDA 12.9 and route the fused backend
